+++ tests/testthat/test-built.R
@@ -15,3 +15,15 @@ test_that("corrupted Build field", {
   desc$set(Built = "foobar")
   expect_error(desc$get_built(), "corrupted")
 })
+
+test_that("get built works with multiples lines", {
+  expect_identical(
+    desc_get_built(test_path("D16")),
+    list(
+      R = R_system_version("3.4.1"),
+      Platform = "x86_64-apple-darwin15.6.0",
+      Date = "2017-09-14 20:30:19 UTC",
+      OStype = "unix"
+    )
+  )
+})