+++ NEWS.md
@@ -1,5 +1,8 @@
 # desc (development version)
 
+* `desc_get_built()` no longer fails if Built spans multiple lines 
+  (#145, @seankross).
+
 * An empty `Depends` field is now properly normalized and formatted.
   (#148, @kevinushey)